pro5-master/pro5-master/myapp/requirements.txt
djangotoolbox
django_mongodb_engine
django-easy-maps
django_extensions
boto3